+++ tests/fixtures/unsupported_syntax_features/unsupported_object_property_unset.php
@@ -1,0 +1,7 @@
+<?php
+class Box {
+    public $name;
+}
+
+$box = new Box();
+unset($box->name);